// src/app/models/voto.ts
export class Voto {
    "id": string;
    "candidato": string;
    "votos": number;
    "designadoPrincipal": string;
    "ram": number;
    "partido": string;
    "color": string
}

export class VotoWithoutID {
    "candidato": string;
    "votos": number;
    "designadoPrincipal": string;
    "ram": number;
    "partido": string;
    "color": string
}